ETCM-1001 Create new package mining and move related code there
