+++ src/main/resources/conf/base.conf
@@ -287,7 +287,7 @@ mantis {
     # Also used by the generic `BlockGenerator`.
     block-cashe-size = 30
 
-    # See io.iohk.ethereum.consensus.Protocol for the available protocols.
+    # See io.iohk.ethereum.consensus.mining.Protocol for the available protocols.
     # Declaring the protocol here means that a more protocol-specific configuration
     # is pulled from the corresponding consensus implementation.
     # For example, in case of ethash, a section named `ethash` is used.
