+++ src/main/scala/io/iohk/ethereum/consensus/mining/MiningBuilder.scala
@@ -1,10 +1,16 @@
-package io.iohk.ethereum.consensus
+package io.iohk.ethereum.consensus.mining
 
-import io.iohk.ethereum.consensus.Protocol.NoAdditionalPoWData
-import io.iohk.ethereum.consensus.Protocol.RestrictedPoWMinerData
+import io.iohk.ethereum.consensus.mining.Protocol.AdditionalPoWProtocolData
+import io.iohk.ethereum.consensus.mining.Protocol.NoAdditionalPoWData
+import io.iohk.ethereum.consensus.mining.Protocol.RestrictedPoWMinerData
+import io.iohk.ethereum.consensus.pow.EthashConfig
 import io.iohk.ethereum.consensus.pow.PoWMining
 import io.iohk.ethereum.consensus.pow.validators.ValidatorsExecutor
-import io.iohk.ethereum.nodebuilder._
+import io.iohk.ethereum.nodebuilder.BlockchainBuilder
+import io.iohk.ethereum.nodebuilder.BlockchainConfigBuilder
+import io.iohk.ethereum.nodebuilder.NodeKeyBuilder
+import io.iohk.ethereum.nodebuilder.StorageBuilder
+import io.iohk.ethereum.nodebuilder.VmBuilder
 import io.iohk.ethereum.utils.Config
 import io.iohk.ethereum.utils.Logger
 
@@ -15,7 +21,7 @@ trait MiningBuilder {
 /** A consensus builder is responsible to instantiate the consensus protocol.
   * This is done dynamically when Mantis boots, based on its configuration.
   *
-  * @see [[io.iohk.ethereum.consensus.Mining Consensus]],
+  * @see [[Mining Consensus]],
   *      [[io.iohk.ethereum.consensus.pow.PoWMining PoWConsensus]],
   */
 trait StdMiningBuilder extends MiningBuilder {
@@ -34,14 +40,14 @@ trait StdMiningBuilder extends MiningBuilder {
 
   //TODO [ETCM-397] refactor configs to avoid possibility of running mocked or
   // restricted-pow consensus on real network like ETC or Mordor
-  protected def buildPoWConsensus(): pow.PoWMining = {
-    val specificConfig = pow.EthashConfig(mantisConfig)
+  protected def buildPoWConsensus(): PoWMining = {
+    val specificConfig = EthashConfig(mantisConfig)
 
     val fullConfig = newConfig(specificConfig)
 
     val validators = ValidatorsExecutor(blockchainConfig, miningConfig.protocol)
 
-    val additionalPoWData = miningConfig.protocol match {
+    val additionalPoWData: AdditionalPoWProtocolData = miningConfig.protocol match {
       case Protocol.PoW | Protocol.MockedPow => NoAdditionalPoWData
       case Protocol.RestrictedPoW            => RestrictedPoWMinerData(nodeKey)
     }
