+++ src/main/scala/io/iohk/ethereum/consensus/pow/validators/OmmersValidator.scala
@@ -2,8 +2,8 @@ package io.iohk.ethereum.consensus.pow.validators
 
 import akka.util.ByteString
 
-import io.iohk.ethereum.consensus.GetBlockHeaderByHash
-import io.iohk.ethereum.consensus.GetNBlocksBack
+import io.iohk.ethereum.consensus.mining.GetBlockHeaderByHash
+import io.iohk.ethereum.consensus.mining.GetNBlocksBack
 import io.iohk.ethereum.consensus.pow.validators.OmmersValidator.OmmersError
 import io.iohk.ethereum.consensus.pow.validators.OmmersValidator.OmmersValid
 import io.iohk.ethereum.consensus.validators.BlockHeaderError
