+++ src/main/scala/io/iohk/ethereum/testmode/TestEthBlockServiceWrapper.scala
@@ -2,7 +2,7 @@ package io.iohk.ethereum.testmode
 
 import akka.util.ByteString
 
-import io.iohk.ethereum.consensus.Mining
+import io.iohk.ethereum.consensus.mining.Mining
 import io.iohk.ethereum.domain.Block
 import io.iohk.ethereum.domain.BlockHeader
 import io.iohk.ethereum.domain.Blockchain
