+++ src/test/scala/io/iohk/ethereum/consensus/mining/MiningConfigs.scala
@@ -1,4 +1,4 @@
-package io.iohk.ethereum.consensus
+package io.iohk.ethereum.consensus.mining
 
 import akka.util.ByteString
